+++ src/model/Model_Taglink.h
@@ -49,7 +49,7 @@ public:
     * Return the Data record pointer for the given payee name
     * Returns 0 when payee not found.
     */
-    Data* get(const wxString& refType, int refId, int tagId);
+    Data* get(const wxString& refType, int64 refId, int64 tagId);
 
     /**
     * Return a map of all tags
@@ -60,12 +60,12 @@ public:
     * Return a map of all tags for the specified transaction
     * Mostly useful to return a sorted list of tagnames associated with a transaction
     */
-    std::map<wxString, int> get(const wxString& refType, int refId);
+    std::map<wxString, int> get(const wxString& refType, int64 refId);
 
     /* Delete all tags for a REFTYPE + REFID */
-    void DeleteAllTags(const wxString& refType, int refID);
+    void DeleteAllTags(const wxString& refType, int64 refID);
 
-    int update(const Data_Set& rows, const wxString& refType, int refId);
+    int update(const Data_Set& rows, const wxString& refType, int64 refId);
 };
 
 #endif //
